move Post to components/
